+++ R/check_hyper_params.R
@@ -18,30 +18,18 @@ check_hyper_params <- function(X_names, params) {
   logger::log_debug("Checking hyper parameters...")
 
   # Input params checks --------------------------------------------------------
-  ntrees_rf <- getElement(params, "ntrees_rf")
-  if (length(ntrees_rf) == 0) {
-    ntrees_rf <- 20
+  ntrees <- getElement(params, "ntrees")
+  if (length(ntrees) == 0) {
+    ntrees <- 20
   } else {
-    if (!inherits(ntrees_rf, "numeric")) {
-      stop("Invalid 'ntrees_rf' input. Please input a number.")
+    if (!inherits(ntrees, "numeric")) {
+      stop("Invalid 'ntrees' input. Please input a positive integer")
+    }
+    if (ntrees<1) {
+      stop("Invalid 'ntrees' input. Please input a positive integer")
     }
   }
-  params[["ntrees_rf"]] <- ntrees_rf
-
-  ntrees_gbm <- getElement(params, "ntrees_gbm")
-  if (length(ntrees_gbm) == 0) {
-    ntrees_gbm <- 20
-  } else {
-    if (!inherits(ntrees_gbm, "numeric")) {
-      stop("Invalid 'ntrees_gbm' input. Please input an integer.")
-    }
-  }
-  params[["ntrees_gbm"]] <- ntrees_gbm
-
-  if (params[["ntrees_gbm"]] + params[["ntrees_rf"]] == 0) {
-    stop("The total number of trees (ntrees_rf + ntrees_gbm) has to be
-         greater than 0")
-  }
+  params[["ntrees"]] <- ntrees
 
   node_size <- getElement(params, "node_size")
   if (length(node_size) == 0) {