+++ tests/testthat/test-generate_rules_matrix.R
@@ -10,8 +10,7 @@ test_that("Rules Extracted Correctly", {
   ite_method <- "bart"
   learner_ps <- "SL.xgboost"
   learner_y <- NA
-  ntrees_rf <- 100
-  ntrees_gbm <- 50
+  ntrees <- 100
   node_size <- 20
   max_nodes <- 5
   max_depth <- 15
@@ -30,7 +29,7 @@ test_that("Rules Extracted Correctly", {
                            learner_y = learner_y)
 
   # Step 3: Generate rules list
-  initial_rules <- generate_rules(X, ite, ntrees_rf, ntrees_gbm, node_size,
+  initial_rules <- generate_rules(X, ite, ntrees, node_size,
                                   max_nodes, max_depth)
 
   rules <- filter_irrelevant_rules(initial_rules, X, ite, t_decay)